Docs: followApi로 follow 관련로직 이동
Resolves: #119
