+++ src/api/followApi.js
@@ -16,3 +16,19 @@ export const getFollow = async (accountname, pageParam, follow, nextPage) => {
 		console.log(error);
 	}
 };
+
+export const postFollow = async (accountname) => {
+	try {
+		const res = await accessInstance.post(`/profile/${accountname}/follow`);
+	} catch (error) {
+		console.error(error);
+	}
+};
+
+export const delUnFollow = async (accountname) => {
+	try {
+		const res = await accessInstance.delete(`/profile/${accountname}/unfollow`);
+	} catch (error) {
+		console.error(error);
+	}
+};
