+++ src/api/profileApi.js
@@ -18,22 +18,6 @@ export const getMyInfo = async () => {
 	}
 };
 
-export const postFollow = async (accountname) => {
-	try {
-		const res = await accessInstance.post(`/profile/${accountname}/follow`);
-	} catch (error) {
-		console.error(error);
-	}
-};
-
-export const delUnFollow = async (accountname) => {
-	try {
-		const res = await accessInstance.delete(`/profile/${accountname}/unfollow`);
-	} catch (error) {
-		console.error(error);
-	}
-};
-
 export const postAccountValid = async (accountname) => {
 	try {
 		const res = await instance.post(`/user/accountnamevalid/`, {
